grust.mapping: Account for GI-only types
These types represent anonymous structs or unions inside structure
definitions. They are created by g-ir-wrapper and don't carry c:type
attributes.

For now, we just emit them with their invented GI names, as those are
suitably underscore-prefixed and made unique among their kind. If one
happens to be a union (as is the one inside GValue), the generated code
is invalid as the containing struct's definition has a member of the
unemitted union type. The best way to deal with it in the near term
is to suppress the emission of the containing struct type (feature
to be delivered) and provide a manual definition of the structure
layout papering over the union with fixed-width types.