Fix up discriminant values of integer constants
If the value of a fixed-width signed integer constant is out of range
for the target type, but within the bit width, convert the value from its
twos-complement representation to avoid warning on "overflowing_literals".

Similar treatment is done for a constant of any unsigned integer type
initialized with a negative value (likely meaning that the GI scanner
reformulated bit reversal as arithmentic negation to keep within the value
syntax). Here, we just emit the value literal with type syntax `i64`
and explicitly convert to the destination type.