grust.output: Context manager classes to manage output files
FileOutput gets a temporary file for output and replaces the target file
once the `with` context has run without exceptions; cleanup on exceptions
removes the temporary file.

DirectOutput is a workalike for standard output which does nothing.
sys.stdout cannot be directly used in `with` because its __exit__
method closes the stream.