Consistently produce UTF-8 output on Python 2.x and 3.x
The output encoding was specified for the project templates,
but not for the custom template object. It was actually wrong
to write bytestring output to a text file on Python 3, so
generation without customized templates was broken, while custom
templates produced output in a default locale-dependent
encoding which may differ from UTF-8.

Now template output is obtained strictly as Unicode and the output
file is written with explicit UTF-8 encoding and system-default
newline conversion. For stdout, however, the default encoding is used
as before.